+++ README.md
@@ -1,10 +1,11 @@
 # Hidden Folder for Obsidian  
 
-Hidden Folder in Obsidian:
+Hidden Folder in Obsidian:  
 
-- Regular expression
-- Ribbon icon for enable or disable
+- Regular expression  
+- Ribbon icon for enable or disable  
+- Translation  
 
 ## Demo
 
-![DEMO](doc/demo.gif)
+![DEMO](doc/demo.gif)
