+++ manifest.json
@@ -1,7 +1,7 @@
 {
 	"id": "hidden-folder-obsidian",
 	"name": "Hidden Folder",
-	"version": "1.0.1",
+	"version": "1.0.3",
 	"minAppVersion": "0.15.0",
 	"description": "Hidden Folder",
 	"author": "ptrsvltns",
