+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "hidden-folder-obsidian",
-	"version": "1.0.1",
+	"version": "1.0.3",
 	"description": "Hidden Folder",
 	"main": "main.js",
 	"scripts": {
