+++ src/global.d.ts
@@ -1,0 +1,5 @@
+export {}
+
+declare global {
+  interface Window { i18next?: { language: string }; }
+}
