+++ src/lang/en.ts
@@ -1,0 +1,2 @@
+export default {
+}
