+++ src/lang/zh.ts
@@ -1,0 +1,18 @@
+export default {
+  "Hidden Folder": "文件隐藏",
+  "Restore display": "恢复显示",
+  "folders": "文件夹",
+  "Folders": "文件夹",
+  "folder": "文件夹",
+  "Hidden display": "隐藏显示",
+  "Show Folders": "显示文件夹",
+  "Hidden Folders": "隐藏文件夹",
+  "Rules": "规则",
+  "Regular expression": "正则表达式",
+  "Example": "例子",
+  "Enable": "开启",
+  "Enable to hidden folder": "是否开启隐藏",
+  "Saving": "保存中",
+  "Settings is saved": "配置已保存",
+  "Save": "保存"
+}
